Avoid indexing anything into SLM history indices when SLM disabled
This commit adds a check to the `SnapshotHistoryStore.putAsync` method
that checks to see whether SLM is enabled prior to indexing the history
document.

Resolves failures from #46508 that were caused by a retention job being
kicked off, the test finishing and deleting all indices (including the
`.slm-history-*` index), and then retention completing and indexing
another history document. This caused the `.slm-history-*` index to be
re-created and the test to fail due to the shard lock check in
`InternalTestCluster.assertAfterTest` that checks all shards being
unlocked.